made the target lock work, so all follow this thing type commnds work again, and added the pose pipeline as default.
