+++ robotbase/src/main/java/frc/robot/Robot.java
@@ -17,7 +17,6 @@ import frc.robot.subsystems.CommandSwerveDrivetrain;
 import frc.robot.subsystems.EndAffector;
 import frc.robot.subsystems.ExtensionArm;
 import frc.robot.subsystems.Intake;
-import frc.robot.subsystems.PhotonVisionCamera;
 import frc.robot.subsystems.Pivot;
 import frc.robot.subsystems.Shooter;
 import frc.robot.subsystems.ShooterPhotonCamera;
@@ -44,7 +43,7 @@ public class Robot extends TimedRobot {
   public static DriverControls driverControls;
   public static CodriverControls codriverControls;
 
-  public static PhotonVisionCamera shooterCam;
+  public static ShooterPhotonCamera shooterCam;
 
   public static EndAffector endAffector;
