--- robotbase/src/main/java/frc/robot/commands/drive/TargetLock.java
@@ -1,33 +1,0 @@
-package frc.robot.commands.drive;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants.SWERVE;
-import frc.robot.Robot;
-import frc.robot.state.RobotState.DriveControlState;
-import frc.robot.subsystems.PhotonVisionCamera;
-
-public class TargetLock extends Command {
-  private int m_pipelineIndex;
-  private PhotonVisionCamera m_camera;
-  private int m_defaultPipeline;
-
-  public TargetLock(int pipelineIndex, PhotonVisionCamera camera, int defaultPipeline) {
-    m_pipelineIndex = pipelineIndex;
-    m_camera = camera;
-    m_defaultPipeline = defaultPipeline;
-  }
-
-  @Override
-  public void initialize() {
-    Robot.state.setDriveControlState(DriveControlState.TARGET_LOCK);
-    m_camera.setPipeline(m_pipelineIndex);
-    SWERVE.TARGET_LOCK_ROTATION_PID_CONTROLLER.reset();
-  }
-
-  @Override
-  public void end(boolean interrupted) {
-    Robot.state.setDriveControlState(DriveControlState.FIELD_RELATIVE);
-    m_camera.setPipeline(m_defaultPipeline);
-    SWERVE.TARGET_LOCK_ROTATION_PID_CONTROLLER.setP(SWERVE.TARGET_LOCK_ROTATION_KP);
-  }
-}
